Add extra menu on card right click and custom thumbnail selection
This menu is to allow for editing metadata of show, such as the title,
thumbnail, ect.

The full implementations of these, with the exception of thumbnail
selection, have not been implemented.

In addition to this, the code for the play buttons of the cards have
been cleaned up and pulled into its own function.