Rework select episode screen to use Rc str
A static reference to anime is created at the top level draw function
(`draw_anime_expand`), which is then passed to other functions reliant
on it.

This has the consequence of needing an unsafe pointer transmute to get
around borrowing `anime_database` and mutating it later in
`draw_episode_list`, which can update `anime` when an episode is
clicked. The only mutation is by `update_watched`, which reassigns
variables in `anime` and will never invalidate the references.