Refactor `json_database` to make it more ergonomic hold references
Previously, the json map would have to clone items from the deserialized
json rather than holding references to them. This greatly increased its
memory footprint and time it took to optimize.

With this commit, an intermediary struct is added that holds both the
raw deserialized json, and the map holding references to it. With this
self-referential struct, however, introduces memory invariances and
several unsafe keywords. This should not be a problem, however, as all
of the call sites will not cause any invalidated references.

This commit also seralizes many of the structs into boxed slices, as
these slices are never to mutate.

Also a `search_map` is introduced among the regular `map`, which
tokenizes by whitespace in preparation for a search feature.

Last thing to note, this commit, I believe, introduces a regression in
deserializing the database. Need to look into this more, but currently
takes 100ms to deserialize in release mode.