Cache anime list and only show animes that exist
Cache anime list rather than having to pull from `anime_map` and sort
everytime `animes()` is called.

Since it is a function called every frame, a reference to the cached
slice is returned instead.

This also allows for more expensive operations to be performed while
building the anime list, such as checking if the directories exist.