Use unsafe dereferencing instead of transmutes and bound lifetimes
Using transmutes can cause a variety of unsafe casting, as demonstrated
in `e7a0425fb366ec813b3649f3539304393910f117`, which was caused by an
inappropriate transmute, when it only really needed to unbound a
lifetime and decouple references.

This commit also attempts to bind multiple references to a `frame`
lifetime, which, as it states, should only live for a frame, and
absolutely should not cross frame boundaries.