String manager and cargo clippy
String manager to avoid creating new strings for every formatted string.
Such a case shows up when displaying titles in the main screen. When a
string exceeds the card width, multiple string combinations are needed
in order to determine the truncation required to keep the string from
overflowing. Once it is calculated, however, it does not need to keep
recomputed every frame, as the same truncation is to be expected for
that string.