Replace http mutex checking with mpsc channels
Rather than polling a shared vector mutex, the main loop polls an mpsc
receiver. Requests are then done via a request function that takes a
sender, and request kind. The function then spawns a tokio thread that
processes the http request and sends the data back to the receiver.

This change brings about +14% performance difference tested in release
mode and using the `-f` flag.